+++ CHANGES.md
@@ -27,5 +27,6 @@ Apollo 2.5.0
 * [Feature: Support exporting and importing configurations for specified applications and clusters](https://github.com/apolloconfig/apollo/pull/5517)
 * [doc: Add rust apollo client link](https://github.com/apolloconfig/apollo/pull/5514)
 * [Perf: optimize namespace-related interface](https://github.com/apolloconfig/apollo/pull/5518)
+* [Perf: Replace synchronized multimap with concurrent hashmap in NotificationControllerV2 for better performance](https://github.com/apolloconfig/apollo/pull/5532)
 ------------------
 All issues and pull requests are [here](https://github.com/apolloconfig/apollo/milestone/16?closed=1)